+++ CHANGELOG.md
@@ -5,8 +5,15 @@ v0.2.0 (in development)
   `Calendar`
 - Gave `MonthShape` `nth_date()` and `dates()` methods for obtaining `Date`
   objects belonging to the month
-- Gave `Date` `pred()` and `succ()` methods for getting the previous & next
-  dates
+- New `Date` methods:
+    - `pred()` — returns the previous date
+    - `succ()` — returns the next date
+    - `later()` — returns an iterator over all later dates
+    - `earlier()` — returns an iterator over all earlier dates
+    - `and_later()` — returns an iterator that yields the receiver and all
+      later dates
+    - `and_earlier()` — returns an iterator that yields the receiver and all
+      earlier dates
 - Added a `Weekday` enum and a `Date::weekday()` method
 
 v0.1.0 (2023-05-02)